refactor: move general response types file
